% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableDT.R
\name{editableDT}
\alias{editableDT}
\title{Server function of editableDT Shiny module}
\usage{
editableDT(
  input,
  output,
  session,
  data,
  length = 50,
  cols = 1:7,
  showButtons = TRUE,
  enableSave = TRUE,
  editable = NULL
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{A reactive data object}

\item{length}{numeric desired length of string}

\item{cols}{numeric Initial columns to display}

\item{showButtons}{logical}

\item{enableSave}{logical}

\item{editable}{logical}
}
\description{
Server function of editableDT Shiny module
}
