% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableDT2.R
\name{editableDT2}
\alias{editableDT2}
\title{Server function of editableDT2 Shiny module}
\usage{
editableDT2(input, output, session, data, length = 50)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{A reactive data object}

\item{length}{numeric desired length}
}
\description{
Server function of editableDT2 Shiny module
}
