% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableDT3.R
\name{editableDT3}
\alias{editableDT3}
\title{Server function of abEditData Shiny module}
\usage{
editableDT3(
  input,
  output,
  session,
  data = reactive(NULL),
  inputwidth = reactive(100),
  mode = reactive(2),
  excludeCols = c(),
  editableRowNames = FALSE,
  defaultlen = 20
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{A data object}

\item{inputwidth}{Numeric indicating default input width in pixel}

\item{mode}{An integer}

\item{excludeCols}{Character A vector of column names in data to hide from the user}

\item{editableRowNames}{Logical If TRUE, user will be allowed to edit row names of data.}
}
\description{
Server function of abEditData Shiny module
}
