% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{editData}
\alias{editData}
\title{A shiny app for editing a 'data.frame'}
\usage{
editData(
  data = NULL,
  viewer = "dialog",
  length = 50,
  cols = 1:7,
  status = "default",
  showButtons = TRUE,
  enableSave = TRUE,
  editable = NULL,
  formatList = NULL,
  ...
)
}
\arguments{
\item{data}{A tibble or a tbl_df or a data.frame to manipulate}

\item{viewer}{Specify where the gadget should be displayed. Possible choices are c("dialog","browser","pane")}

\item{length}{Numeric desired maximum length of string}

\item{cols}{numeric}

\item{status}{character. dropdownButton status. One of c("default","info","primary","danger","warning","success")}

\item{showButtons}{logical}

\item{enableSave}{logical}

\item{editable}{logical}

\item{formatList}{Null or list. Format list to be passed to formatStyle}

\item{...}{Further arguments to be passed to datatable()}
}
\value{
A manipulated 'data.frame' or NULL
}
\description{
A shiny app for editing a 'data.frame'
}
\examples{
library(shiny)
library(editData)
# Only run examples in interactive R sessions
if (interactive()) {
    result<-editData(mtcars)
    result
}
}
