% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData3.R
\name{editData3}
\alias{editData3}
\title{A shiny app for editing a 'data.frame'}
\usage{
editData3(
  data = NULL,
  viewer = c("dialog", "browser", "pane")[1],
  mode = c(1:4)[4]
)
}
\arguments{
\item{data}{A tibble or a tbl_df or a data.frame to manipulate}

\item{viewer}{Specify where the gadget should be displayed. Possible choices are c("dialog","browser","pane")}

\item{mode}{An integer}
}
\value{
A manipulated 'data.frame' or NULL
}
\description{
A shiny app for editing a 'data.frame'
}
