% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editFiles.R
\name{file2ext}
\alias{file2ext}
\title{Extract extension from a file name}
\usage{
file2ext(filename)
}
\arguments{
\item{filename}{A character string naming a file}
}
\description{
Extract extension from a file name
}
