% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerInput3.R
\name{integerInput3}
\alias{integerInput3}
\title{Create a side-by-side integerInput}
\usage{
integerInput3(
  inputId,
  label,
  value,
  min = NA,
  max = NA,
  step = NA,
  width = 100,
  ...
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}

\item{width}{The width of the input in pixel}

\item{...}{arguments to be passed to integerInput}

\item{title}{integerInput3}
}
\description{
Create a side-by-side integerInput
}
\examples{
library(shiny)
# Only run examples in interactive R sessions
if (interactive()) {
  ui <- fluidPage(
         textInput3("id", "id", ""),
         integerInput3("score","score",value=1)
    )
    server <- function(input, output) {

    }
    shinyApp(ui, server)
}
}
