% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableDT2.R
\name{makeShort}
\alias{makeShort}
\title{Truncate string to desired length}
\usage{
makeShort(x, length = 50)
}
\arguments{
\item{x}{A vector}

\item{length}{numeric desired string length}
}
\description{
Truncate string to desired length
}
