% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myget.R
\name{myget}
\alias{myget}
\title{Return the Value of a Named data.frame}
\usage{
myget(x)
}
\arguments{
\item{x}{Name of data.frame}
}
\description{
Return the Value of a Named data.frame
}
\examples{
myget("iris")
myget("mtcars")
}
