% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editFiles.R
\name{myimport}
\alias{myimport}
\title{Read in a data.frame from a file}
\usage{
myimport(file, ...)
}
\arguments{
\item{file}{A character string naming a file}

\item{...}{Further arguments to be passed to rio::import}
}
\description{
Read in a data.frame from a file
}
