% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editFiles.R
\name{myimport_csv}
\alias{myimport_csv}
\title{read csv file}
\usage{
myimport_csv(file, ...)
}
\arguments{
\item{file}{A character string naming a file}

\item{...}{Further arguments to be passed to read.csv}
}
\description{
read csv file
}
