% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInput3.R
\name{numericInput3}
\alias{numericInput3}
\title{Create a side-by-side numericInput}
\usage{
numericInput3(
  inputId,
  label,
  value,
  min = NA,
  max = NA,
  step = NA,
  width = 100,
  ...
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}

\item{width}{The width of the input in pixel}

\item{...}{arguments to be passed to numericInput}
}
\description{
Create a side-by-side numericInput
}
\examples{
library(shiny)
# Only run examples in interactive R sessions
if (interactive()) {
  ui <- fluidPage(
         textInput3("id", "id", ""),
         numericInput3("score","score",value=1)
    )
    server <- function(input, output) {

    }
    shinyApp(ui, server)
}
}
