% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInput3.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Sample Data for testing 'editData' addin}
\format{A data.frame with 4 rows and 6 variables:
\describe{
\item{name}{Last name}
\item{age}{age in years}
\item{country}{Country Name}
\item{sex}{sex, A factor with two levels.}
\item{bloodType}{Blood Type. A factor with four levels}
\item{date}{Date}
}}
\usage{
sampleData
}
\description{
A sample dataset containing data for 4 people
}
\keyword{datasets}
