% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInput3.R
\name{selectInput3}
\alias{selectInput3}
\title{Create a side-by-side selectInput}
\usage{
selectInput3(..., width = 100)
}
\arguments{
\item{...}{arguments to be passed to selectInput}

\item{width}{The width of the input in pixel}
}
\description{
Create a side-by-side selectInput
}
\examples{
library(shiny)
# Only run examples in interactive R sessions
if (interactive()) {
  ui <- fluidPage(
         selectInput3("sex", "sex", choices=c("Male","Female")),
         selectInput3("smoking", "smokingStatus", choices=c("Never","Ex-smoker","Smoker"))
    )
    server <- function(input, output) {

    }
    shinyApp(ui, server)
}
}
