% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInput3.R
\name{textInput3}
\alias{textInput3}
\title{Create a side-by-side textInput control for entry of unstructured text values}
\usage{
textInput3(inputId, label, value = "", width = 100, bg = NULL, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input in pixel}

\item{bg}{The color of text}

\item{...}{arguments to be passed to textInput}
}
\description{
Create a side-by-side textInput control for entry of unstructured text values
}
\examples{
library(shiny)
# Only run examples in interactive R sessions
if (interactive()) {
  ui <- fluidPage(
         textInput3("id", "id", ""),
         textInput3("name","name","")
    )
    server <- function(input, output) {

    }
    shinyApp(ui, server)
}
}
