% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableDT2.R
\name{editableDT}
\alias{editableDT}
\title{Server function of editData Shiny module}
\usage{
editableDT(input, output, session, dataname = reactive(""),
  data = reactive(NULL), inputwidth = reactive(100))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{dataname}{A string of representing data name}

\item{data}{A data object}

\item{inputwidth}{Numeric indicating default input width in pixel}
}
\description{
Server function of editData Shiny module
}
