% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableDT.R
\name{editableDT}
\alias{editableDT}
\title{Server function of editableDT Shiny module}
\usage{
editableDT(input, output, session, data)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{A reactive data object}
}
\description{
Server function of editableDT Shiny module
}
