% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editableDT2.R
\name{editableDTUI}
\alias{editableDTUI}
\title{UI of editData Shiny module}
\usage{
editableDTUI(id)
}
\arguments{
\item{id}{A string}
}
\description{
UI of editData Shiny module
}
\examples{
library(shiny)
# Only run examples in interactive R sessions
if (interactive()) {
  ui <- fluidPage(
      textInput("mydata","Enter data name",value="sampleData"),
      editableDTUI("table1"),
      verbatimTextOutput("test"),
      editableDTUI("table2"),
      verbatimTextOutput("test2")
  )
  server <- function(input, output) {
     df=callModule(editableDT,"table1",data=reactive(mydf()),inputwidth=reactive(170))

     output$test=renderPrint({
        str(df())
     })
    mydf<-editData::sampleData
    df2=callModule(editableDT,"table2",data=reactive(mydf))
    output$test2=renderPrint({
       str(df2())
    })
 }
 shinyApp(ui, server)
}
}
