% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{editData}
\alias{editData}
\title{A shiny app for editing a 'data.frame'}
\usage{
editData(data = NULL, viewer = "dialog", mode = 2)
}
\arguments{
\item{data}{A tibble or a tbl_df or a data.frame to manipulate}

\item{viewer}{Specify where the gadget should be displayed. Possible choices are c("dialog","browser","pane")}

\item{mode}{An integer}
}
\value{
A manipulated 'data.frame' or NULL
}
\description{
A shiny app for editing a 'data.frame'
}
\examples{
library(shiny)
library(editData)
# Only run examples in interactive R sessions
if (interactive()) {
    result<-editData(mtcars)
    result
}
}
